Ampliado tamaño del campo puntos de tinyint a int. Ya si que muestra puntos por encima de 127...

+++ servidor/crearbd.sql
@@ -37,7 +37,7 @@ create table cartas (
    ataque      tinyint	      not null,
    defensa     tinyint	      not null,
    coste       tinyint	      not null,
-   puntos      tinyint	      not null,
+   puntos      int            not null,
    nivel       tinyint	      not null,
    vida	       tinyint	      not null,
    habilidades text	      not null,
